requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
